requirements.txt
python-dotenv
Telethon
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
dbt-postgres
ultralytics
Pillow